+++ src/Logger.php
@@ -4,6 +4,7 @@ namespace Seablast\Logger;
 
 use Psr\Log\AbstractLogger;
 use Psr\Log\LoggerInterface;
+use Psr\Log\LogLevel;
 use Seablast\Logger\LoggerTime;
 
 /**
@@ -11,11 +12,22 @@ use Seablast\Logger\LoggerTime;
  * - add $RUNNING_TIME = $this->getLastRunningTime();
  * - KEEP dieGraciously
  * - only log calls Seablast\Logger\Logger and catches ErrorLogFailureException('error_log() => return false
- * - when adding PHP/8 support, add :void to the inherited methods and remove PHP/5 support
+ * TODO when adding PHP/8 support, add :void to the inherited methods and remove PHP/5 support
  */
 class Logger extends AbstractLogger implements LoggerInterface
 {
-// phpcs:disable Generic.Files.LineLength
+    // Define constants for configuration keys
+    // phpcs:disable PSR12.Properties.ConstantVisibility.NotFound
+    // todo remove phpcs exception when PHP/5 support removed
+    const CONF_ERROR_LOG_MESSAGE_TYPE = 'error_log_message_type';
+    const CONF_LOGGING_FILE = 'logging_file';
+    const CONF_LOGGING_LEVEL = 'logging_level';
+    const CONF_LOGGING_LEVEL_NAME = 'logging_level_name';
+    const CONF_LOGGING_LEVEL_PAGE_SPEED = 'logging_level_page_speed';
+    const CONF_LOG_MONTHLY_ROTATION = 'log_monthly_rotation';
+    const CONF_LOG_PROFILING_STEP = 'log_profiling_step';
+    const CONF_MAIL_FOR_ADMIN_ENABLED = 'mail_for_admin_enabled';
+    // phpcs:enable
 
     /** @var array<mixed> int,string,bool,array */
     protected $conf = array();
@@ -40,13 +52,13 @@ class Logger extends AbstractLogger implements LoggerInterface
             array( // default values
                 // 0 = send message to PHP's system logger;
                 // recommended is however 3, i.e. append to the file destination set in the field 'logging_file'
-                'error_log_message_type' => 0,
+                self::CONF_ERROR_LOG_MESSAGE_TYPE => 0,
                 // if error_log_message_type equals 3, the message is appended to this file destination (path and name)
-                'logging_file' => '',
-                // log up to the level set here, default=5 = debug
-                'logging_level' => 5,
+                self::CONF_LOGGING_FILE => '',
+                // verbosity: log up to the level set here, default=5 = debug
+                self::CONF_LOGGING_LEVEL => 5,
                 // rename or renumber, if needed
-                'logging_level_name' => array(
+                self::CONF_LOGGING_LEVEL_NAME => array(
                     0 => 'unknown',
                     1 => 'fatal',
                     'error',
@@ -56,29 +68,30 @@ class Logger extends AbstractLogger implements LoggerInterface
                     'speed'
                 ),
                 // the logging level to which the page generation speed (i.e. error_number 6) is to be logged
-                'logging_level_page_speed' => 5,
+                self::CONF_LOGGING_LEVEL_PAGE_SPEED => 5,
                 // false => use logging_file with log extension as destination
                 // true => adds .Y-m.log to the logging file
-                'log_monthly_rotation' => true,
+                self::CONF_LOG_MONTHLY_ROTATION => true,
                 // prefix message that took longer than profiling step (float) sec from the previous one by SLOWSTEP
-                'log_profiling_step' => false,
-                // UNCOMMENT only if needed //'log_standard_output' => false, //true, pokud má zároveň vypisovat na obrazovku; false, pokud má vypisovat jen do logu
+                self::CONF_LOG_PROFILING_STEP => false,
+                // UNCOMMENT only if needed //'log_standard_output' => false,
+                // //true, pokud má zároveň vypisovat na obrazovku; false, pokud má vypisovat jen do logu
                 // fatal error may just be written in log,
                 // on production, it is however recommended to set an e-mail, where to announce fatal errors
-                'mail_for_admin_enabled' => false,
+                self::CONF_MAIL_FOR_ADMIN_ENABLED => false,
             ),
             $conf
         );
-        if (!is_int($this->conf['logging_level'])) {
-            throw new \Psr\Log\InvalidArgumentException('The logging_level is not an integer.');
+        if (!is_int($this->conf[self::CONF_LOGGING_LEVEL])) {
+            throw new \Psr\Log\InvalidArgumentException('The logging_level MUST be an integer.');
         }
-        $this->overrideLoggingLevel = $this->conf['logging_level'];
-        //@todo do not use $this->conf but set the class properties right here accordingly; and also provide means to set the values otherwise later
-        //240709 set later is probably not necessary
+        $this->overrideLoggingLevel = $this->conf[self::CONF_LOGGING_LEVEL];
+        //@todo replace $this->conf by the class properties
     }
 
     /**
-     * Class doesn't automatically use any GET parameter to override the set logging level, as it could be used to flood the error log.
+     * Class doesn't automatically use any GET parameter to override the set logging level,
+     * as it could be used to flood the error log.
      * It is however possible to programmatically raise the logging level set in configuration.
      *
      * @param int $newLevel
@@ -218,12 +231,13 @@ class Logger extends AbstractLogger implements LoggerInterface
         $this->log(5, $message, $context);
     }
 
+    // phpcs:disable Generic.Files.LineLength
     /**
      * Error_log() modified to log necessary debug information by application to its own log.
-     * Logs with an arbitrary level, i.e. may not log debug info on production.
+     * Logs with an arbitrary verbosity level, e.g. debug info on production may be omitted.
      * Compliant with PSR-3 http://www.php-fig.org/psr/psr-3/
      *
-     * @param int $level Error level
+     * @param mixed $level int|string Error level
      * @param string $message Message to be logged
      * @param array<int> $context OPTIONAL To enable error log filtering 'error_number' field expected or the first element element expected containing number of error category
      *
@@ -257,9 +271,31 @@ class Logger extends AbstractLogger implements LoggerInterface
         //TODO: přidat proměnnou $line - mělo by být vždy voláno jako basename(__FILE__)."#".__LINE__ , takže bude jasné, ze které řádky source souboru to bylo voláno
         // Ve výsledku do logu zapíše:
         //[Timestamp: d-M-Y H:i:s] [Logging level] [$error_number] [$_SERVER['SCRIPT_FILENAME']] [username@gethostbyaddr($_SERVER['REMOTE_ADDR'])] [sec od startu stránky] $message
-
         if (!is_string($message)) {
-            error_log("wrong message: Logger->log({$level}," . print_r($message, true) . ")");
+            $message = "wrong message type " . gettype($message) . ": Logger->log(" . print_r($level, true) . "," . print_r($message, true) . ")";
+            $this->error($message);
+        }
+        // psr log levels to numbered severity
+        $psr2int = array(
+            LogLevel::EMERGENCY => 0,
+            LogLevel::ALERT     => 1,
+            LogLevel::CRITICAL  => 1,
+            LogLevel::ERROR     => 2,
+            LogLevel::WARNING   => 3,
+            LogLevel::NOTICE    => 4,
+            LogLevel::INFO      => 4,
+            LogLevel::DEBUG     => 5,
+        );
+        if (is_string($level)) {
+            if (array_key_exists($level, $psr2int)) {
+                $level = $psr2int[$level];
+            } else {
+                $this->error('level has unexpected string value ' . $level . ' message: ' . $message);
+                $level = 0;
+            }
+        } elseif (!is_int($level)) {
+            $this->error('level has unexpected type ' . gettype($level) . ' message: ' . $message);
+            $level = 0;
         }
 
         // if context array is set then get the value of the 'error_number' field or the first element
@@ -274,19 +310,17 @@ class Logger extends AbstractLogger implements LoggerInterface
             (
                 $level <= max(
                     array(
-                        $this->conf['logging_level'],
+                        $this->conf[self::CONF_LOGGING_LEVEL],
                         $this->overrideLoggingLevel,
-                        // $this->conf['error_hack_from_get'], //set potentially as GET parameter
-                        //  $ERROR_HACK, //set as variable in the application script
                     )
                 )
             )
             // or log page_speed everytime error_number equals 6 and
             // logging_level_page_speed has at least the severity of logging_level
-            || (($error_number === 6) && ($this->conf['logging_level_page_speed'] <= $this->conf['logging_level']))
+            || (($error_number === 6) && ($this->conf[self::CONF_LOGGING_LEVEL_PAGE_SPEED] <= $this->conf[self::CONF_LOGGING_LEVEL]))
         ) {
             $RUNNING_TIME_PREVIOUS = $this->runningTime;
-            if (((($this->runningTime = round($this->time->getmicrotime() - $this->time->getPageTimestamp(), 4)) - $RUNNING_TIME_PREVIOUS) > $this->conf['log_profiling_step']) && $this->conf['log_profiling_step']) {
+            if (((($this->runningTime = round($this->time->getmicrotime() - $this->time->getPageTimestamp(), 4)) - $RUNNING_TIME_PREVIOUS) > $this->conf[self::CONF_LOG_PROFILING_STEP]) && $this->conf[self::CONF_LOG_PROFILING_STEP]) {
                 $message = "SLOWSTEP " . $message; //110812, PROFILING
             }
 
@@ -295,30 +329,32 @@ class Logger extends AbstractLogger implements LoggerInterface
             //    echo((($level <= 2) ? "<b>" : "") . "{$message} [{$this->runningTime}]" . (($level <= 2) ? "</b>" : "") . "<hr/>" . PHP_EOL); //110811, if fatal or error then bold//111119, RUNNING_TIME
             //}
 
-            $message_prefix = "[" . date("d-M-Y H:i:s") . "] [" . $this->conf['logging_level_name'][$level] . "] [" . $error_number . "] [" . $_SERVER['SCRIPT_FILENAME'] . "] ["
+            $message_prefix = "[" . date("d-M-Y H:i:s") . "] [" . $this->conf[self::CONF_LOGGING_LEVEL_NAME][$level] . "] [" . $error_number . "] [" . $_SERVER['SCRIPT_FILENAME'] . "] ["
                 . $this->user . "@"
                 . (isset($_SERVER['REMOTE_ADDR']) ? gethostbyaddr($_SERVER['REMOTE_ADDR']) : '-')//PHPUnit test (CLI) does not set REMOTE_ADDR
                 . "] [" . $this->runningTime . "] ["
                 . (isset($_SERVER["REQUEST_URI"]) ? $_SERVER["REQUEST_URI"] : '-')//PHPUnit test (CLI) does not set REQUEST_URI
                 . "] ";
             //gethostbyaddr($_SERVER['REMOTE_ADDR'])// co udělá s IP, která nelze přeložit? nebylo by lepší logovat přímo IP?
-            if (($this->conf['error_log_message_type'] == 3) && !$this->conf['logging_file']) {// $logging_file not set and it should be
+            if (($this->conf[self::CONF_ERROR_LOG_MESSAGE_TYPE] == 3) && !$this->conf[self::CONF_LOGGING_FILE]) {// $logging_file not set and it should be
                 $result = error_log($message_prefix . "(error: logging_file should be set!) $message"); // so write into the default destination
                 //zaroven by mohlo poslat mail nebo tak neco .. vypis na obrazovku je asi az krajni reseni
             } else {
-                $messageType = ($this->conf['error_log_message_type'] == 0) ? $this->conf['error_log_message_type'] : 3;
-                $result = ($this->conf['log_monthly_rotation']) ? error_log($message_prefix . $message . (($messageType != 0) ? (PHP_EOL) : ('')), $messageType, "{$this->conf['logging_file']}." . date("Y-m") . ".log") //writes into a monthly rotating file
-                    : error_log($message_prefix . $message . PHP_EOL, $messageType, "{$this->conf['logging_file']}.log"); //writes into one file
+                $messageType = ($this->conf[self::CONF_ERROR_LOG_MESSAGE_TYPE] == 0) ? $this->conf[self::CONF_ERROR_LOG_MESSAGE_TYPE] : 3;
+                $result = ($this->conf[self::CONF_LOG_MONTHLY_ROTATION])
+                    ? error_log($message_prefix . $message . (($messageType != 0) ? (PHP_EOL) : ('')), $messageType, "{$this->conf[self::CONF_LOGGING_FILE]}." . date("Y-m") . ".log") //writes into a monthly rotating file
+                    : error_log($message_prefix . $message . PHP_EOL, $messageType, "{$this->conf[self::CONF_LOGGING_FILE]}.log"); //writes into one file
             }
             // mailto admin. 'mail_for_admin_enabled' has to be an email
-            if ($level == 1 && $this->conf['mail_for_admin_enabled']) {
-                error_log($message_prefix . $message . PHP_EOL, 1, $this->conf['mail_for_admin_enabled']);
+            if ($level == 1 && $this->conf[self::CONF_MAIL_FOR_ADMIN_ENABLED]) {
+                error_log($message_prefix . $message . PHP_EOL, 1, $this->conf[self::CONF_MAIL_FOR_ADMIN_ENABLED]);
             }
         }
         if ($result === false) {
             throw new ErrorLogFailureException('error_log() failed');
         }
     }
+    // phpcs:enable
     /** Alternative way:
       Logging levels
       Log level   Description                                                                       Set bit
@@ -336,5 +372,4 @@ class Logger extends AbstractLogger implements LoggerInterface
       4         00000100  Warnings and Trace
       7         00000111  Warnings, Debug, Information and Trace
      */
-    // phpcs:enable
 }